+++ src/components/ui/button/Button.tsx
@@ -1,35 +1,43 @@
-import React from "react";
+import { borderPxToRem, dimensionsPxToRem, fontPxToRem, pxToRem, styled } from "../../../theme.js";
 
-import classnames from "classnames";
+/**
+ * Call-to-action button.
+ */
+const Button = styled("button", {
+    "height": dimensionsPxToRem(118),
+    "background": "$green1",
+    "border": `${borderPxToRem(1)} solid $green1Darker`,
+    "color": "white",
+    "borderRadius": "1000px",
+    "display": "flex",
+    "alignItems": "center",
+    "justifyContent": "center",
+    "width": "100%",
+    "fontSize": fontPxToRem(26),
+    "padding": "0 1em",
+    "gap": pxToRem(10),
 
-import { makeVariants } from "../../../utils/index.js";
+    "&:disabled": {
+        opacity: 0.5,
+    },
 
-import styles from "./Button.module.scss";
+    "variants": {
+        inline: {
+            true: {
+                width: "auto",
+                display: "inline-flex",
+            },
+        },
+        outline: {
+            true: {
+                background: "white",
+                borderColor: "$buttonBorder",
+                color: "$grey1",
+            },
+        },
+    },
+});
 
-type Variant = "inline" | "outline";
-
-interface Props {
-    variant?: Variant | Variant[];
-}
-
-const Button: React.FC<React.ButtonHTMLAttributes<HTMLButtonElement> & Props> = ({
-    className, children, variant, ...props
-}) => {
-    const variants = makeVariants(variant);
-
-    const cls = classnames(styles.btn, {
-        [styles["btn--inline"]]: variants.includes("inline"),
-        [styles["btn--outline"]]: variants.includes("outline"),
-    }, className);
-
-    return (
-        <button
-            {...props}
-            className={cls}
-            disabled={props.disabled}
-        >{children}
-        </button>
-    );
+export {
+    Button,
 };
-
-export { Button };